Make correct navigation item be selected in header
I've added some JavaScript and a small conditional to correct which navigation item is selected for manually added navigation items. This significantly improves the look and usability for manually added navigation items, and is especially useful if adding nav items for individual entries (a homepage, for example).

The preg_replace is necessary to be able to correctly handle both absolute and relative URLs as inputted by the user.

Unfortunately this does have to be done with JavaScript injection right now as the point at which nav items are decided if they are "selected" or not happens after the modifyCpNav hook that is available to plugins.